refactor: move notifer interface to consumer root
