--- consumers/notifiers/doc.go
@@ -1,6 +1,0 @@
-// Copyright (c) Abstract Machines
-// SPDX-License-Identifier: Apache-2.0
-
-// Package notifiers contain the interface definitions needed to support
-// SuperMQ notifications functionality.
-package notifiers
